Break up the gluestack-ui/util/index.ts file into domain-grouped files
