+++ .vscode/settings.json
@@ -9,20 +9,20 @@
     // Standard className patterns
     ["className\\s*=\\s*[\"'`]([^\"'`]*)[\"'`]", "([^\"'`]*)"],
     ["class\\s*=\\s*[\"'`]([^\"'`]*)[\"'`]", "([^\"'`]*)"],
-    
+
     // TVA patterns (your tva() calls)
     ["tva\\(\\s*{[^}]*base:\\s*[\"'`]([^\"'`]*)[\"'`]", "([^\"'`]*)"],
     ["[\"'`]([^\"'`]*)[\"'`]\\s*:", "([^\"'`]*)"],
-    
+
     // Class utility functions
     ["cn\\(([^)]*)\\)", "\"([^\"]*)\""],
     ["classNames\\(([^)]*)\\)", "\"([^\"]*)\""],
     ["clsx\\(([^)]*)\\)", "\"([^\"]*)\""],
-    
+
     // Gluestack style functions
     ["Style\\(\\s*{[^}]*class:\\s*[\"'`]([^\"'`]*)[\"'`]", "([^\"'`]*)"],
     ["class:\\s*[\"'`]([^\"'`]*)[\"'`]", "([^\"'`]*)"],
-    
+
     // Template literals
     ["className\\s*=\\s*`([^`]*)`", "([^`]*)"],
     ["class\\s*=\\s*`([^`]*)`", "([^`]*)"]
@@ -45,14 +45,18 @@
   "tailwindCSS.lint.invalidTailwindDirective": "error",
   "tailwindCSS.lint.recommendedVariantOrder": "warning",
   "tailwindCSS.classAttributes": [
-    "class", 
-    "className", 
-    "class:list", 
-    "classList", 
+    "class",
+    "className",
+    "class:list",
+    "classList",
     "ngClass"
   ],
   "editor.quickSuggestions": {
     "strings": "on"
   },
-  "editor.autoClosingQuotes": "always"
+  "editor.autoClosingQuotes": "always",
+  "editor.defaultFormatter": "esbenp.prettier-vscode",
+  "[typescript]": {
+    "editor.defaultFormatter": "esbenp.prettier-vscode"
+  }
 }
