+++ packages/gluestack-ui/src/config.ts
@@ -3,12 +3,9 @@ const config = {
   gluestackDir: '.gluestack/cache/gluestack-ui',
   componentsResourcePath: 'src/components/ui',
   nativeWindRootPath: 'nativewind',
-  expoProject: 'expo',
-  nextJsProject: 'nextjs',
-  reactNativeCLIProject: 'react-native-cli',
   tailwindConfigRootPath: 'src/gluestack-ui/templates/tailwind.config.js',
   writableComponentsPath: 'components/ui',
-  branchName: 'feat/v3',  // branch name for the gluestack-ui repo to be used for the CLI
+  branchName: 'feat/v3', // branch name for the gluestack-ui repo to be used for the CLI
   style: 'nativewind',
   providerComponent: 'gluestack-ui-provider',
   gluestackUIPattern: '@/components/ui/',
