+++ packages/gluestack-ui/src/util/clone-repository-at-root.ts
@@ -1,0 +1,89 @@
+import os from 'os';
+import fs from 'fs-extra';
+import simpleGit from 'simple-git';
+import { config } from '../config';
+import { join } from 'path';
+import { log, spinner } from '@clack/prompts';
+import { checkIfFolderExists } from './file-operations';
+
+const homeDir = os.homedir();
+
+const cloneComponentRepo = async (
+  targetPath: string,
+  gitURL: string
+): Promise<void> => {
+  const git = simpleGit();
+  const s = spinner();
+  s.start('⏳ Cloning repository...');
+  try {
+    await git.clone(gitURL, targetPath, [
+      '--depth=1',
+      '--branch',
+      config.branchName,
+    ]);
+    s.stop('\x1b[32m' + 'Cloning successful.' + '\x1b[0m');
+  } catch (err) {
+    s.stop('\x1b[31m' + 'Cloning failed' + '\x1b[0m');
+    throw new Error((err as Error).message);
+  }
+};
+
+const pullComponentRepo = async (targetpath: string): Promise<void> => {
+  const s = spinner();
+  s.start('⏳ Pulling latest changes...');
+  let retry = 0;
+  let success = false;
+  while (!success && retry < 3) {
+    try {
+      await wait(1000);
+      await tryGitPull(targetpath);
+      success = true;
+    } catch (err) {
+      log.error(`\x1b[31mError: ${(err as Error).message}\x1b[0m`);
+      log.error(
+        `\x1b[31mPulling failed - retrying... (Attempt ${retry + 1})\x1b[0m`
+      );
+      retry++;
+    }
+  }
+  if (!success) {
+    s.stop('\x1b[31m' + 'Pulling failed!' + '\x1b[0m');
+    throw new Error('Error pulling remote branch!');
+  } else s.stop('Git pull successful.');
+};
+
+const tryGitPull = async (targetPath: string): Promise<void> => {
+  const git = simpleGit(targetPath);
+  if (fs.existsSync(targetPath)) {
+    await git.pull('origin', config.branchName);
+  } else log.error('\x1b[31m' + 'Target path does not exist' + '\x1b[0m');
+};
+
+const wait = (msec: number): Promise<void> =>
+  new Promise<void>((resolve, _) => {
+    setTimeout(resolve, msec);
+  });
+
+export const cloneRepositoryAtRoot = async (rootPath: string) => {
+  try {
+    const clonedRepoExists = await checkIfFolderExists(rootPath);
+    if (clonedRepoExists) {
+      const git = simpleGit(rootPath);
+      const currBranch = await git.branchLocal();
+      if (currBranch.current !== config.branchName) {
+        fs.removeSync(rootPath);
+        await cloneComponentRepo(rootPath, config.repoUrl);
+      }
+
+      if (currBranch.current === config.branchName) {
+        log.step('Repository already cloned.');
+        await pullComponentRepo(join(homeDir, config.gluestackDir));
+      }
+    } else {
+      await cloneComponentRepo(rootPath, config.repoUrl);
+    }
+  } catch (err) {
+    log.error(`\x1b[31m Cloning failed.\x1b[0m`);
+    throw new Error((err as Error).message);
+  }
+};
